rename files to match tasks carried by enclosed functions
