+++ src/index.js
@@ -1,4 +1,4 @@
-import render from './scripts/dom-handlers/dom';
+import render from './scripts/dom-handlers/render';
 import './style.css';
 
 render();
