+++ src/scripts/components/headerComponent.js
