+++ src/scripts/components/modalComponent.js
